pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-analysis"
version = "0.1.0"
description = "Scientific paper POC: figure vision extraction + text/native-table pipeline with LLM candidates"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "pydantic>=2.6.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[project.scripts]
paper-analysis = "paper_analysis.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
